refactor: 채팅 메시지 관련 로직 제외 모두 삭제
