--- src/main/java/org/example/backendchat/board/BoardRepository.java
@@ -1,9 +1,0 @@
-package org.example.backendchat.board;
-
-import org.example.backendchat.entity.Board;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BoardRepository extends JpaRepository<Board, Long> {
-}
