--- src/main/java/org/example/backendchat/chat/dto/EncryptedChatMessageDTO.java
@@ -1,15 +1,0 @@
-package org.example.backendchat.chat.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class EncryptedChatMessageDTO {
-
-	private String message;
-	private String senderId;
-	private String chatRoomType;
-	private String chatRoomId;
-
-}
