+++ src/main/java/org/example/backendchat/chat/dto/etc/ChatMessageDTO.java
@@ -1,13 +1,13 @@
-package org.example.backendchat.chatroom.dto.etc;
+package org.example.backendchat.chat.dto.etc;
 
 import org.example.backendchat.entity.ChatRoomType;
 
-import lombok.Builder;
 import lombok.Getter;
+import lombok.Setter;
 
 @Getter
-@Builder
-public class ChatRoomChatMessageDTO {
+@Setter
+public class ChatMessageDTO {
 
 	private String message;
 	private Long senderId;
