--- src/main/java/org/example/backendchat/chatroom/dto/req/EnterOneOnOneChatRoomReqDTO.java
@@ -1,13 +1,0 @@
-package org.example.backendchat.chatroom.dto.req;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class EnterOneOnOneChatRoomReqDTO {
-
-    private Long userId;
-    private Long contactUserId;
-
-}
