--- src/main/java/org/example/backendchat/chatroom/dto/res/EnterOneOnOneChatRoomResDTO.java
@@ -1,14 +1,0 @@
-package org.example.backendchat.chatroom.dto.res;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class EnterOneOnOneChatRoomResDTO {
-
-    private Long chatRoomId;
-    private Long participantId1;
-    private Long participantId2;
-
-}
