--- src/main/java/org/example/backendchat/chatroom/dto/res/ExitGroupChatRoomResDTO.java
@@ -1,13 +1,0 @@
-package org.example.backendchat.chatroom.dto.res;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ExitGroupChatRoomResDTO {
-
-    private Long groupChatRoomId;
-    private Long userId;
-
-}
