--- src/main/java/org/example/backendchat/chatroom/dto/res/ExitOneOnOneChatRoomResDTO.java
@@ -1,13 +1,0 @@
-package org.example.backendchat.chatroom.dto.res;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ExitOneOnOneChatRoomResDTO {
-
-    private Long oneOnOneChatRoomId;
-    private Long userId;
-
-}
