--- src/main/java/org/example/backendchat/common/constants/JwtConstants.java
@@ -1,16 +1,0 @@
-package org.example.backendchat.common.constants;
-
-import lombok.Getter;
-
-@Getter
-public enum JwtConstants {
-
-    ACCESS_TOKEN("access"),
-    REFRESH_TOKEN("refresh");
-
-    private String message;
-
-    JwtConstants(String message) {
-        this.message = message;
-    }
-}
