+++ src/main/java/org/example/backendchat/redis/RedisSubscriber.java
@@ -2,10 +2,8 @@ package org.example.backendchat.redis;
 
 import static org.example.backendchat.common.error.ErrorCode.*;
 
-import org.example.backendchat.chat.dto.ChatMessageDTO;
-import org.example.backendchat.chat.dto.EncryptedChatMessageDTO;
+import org.example.backendchat.chat.dto.etc.ChatMessageDTO;
 import org.example.backendchat.common.exception.ChatMessageSendingException;
-import org.example.backendchat.common.utils.MessageCryptUtils;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.connection.MessageListener;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -25,15 +23,15 @@ public class RedisSubscriber implements MessageListener {
 	private final SimpMessageSendingOperations messageTemplate;
 	private final ObjectMapper objectMapper;
 	private final RedisTemplate<String, Object> redisTemplate;
-	private final MessageCryptUtils messageCryptUtils;
 
 	@Override
 	public void onMessage(Message message, byte[] pattern) {
 		try {
 			String publishedMessage = redisTemplate.getStringSerializer().deserialize(message.getBody());
-			EncryptedChatMessageDTO encryptedChatMessageDTO = objectMapper.readValue(publishedMessage, EncryptedChatMessageDTO.class);
-			ChatMessageDTO chatMessageDTO = messageCryptUtils.decryptChatMessage(encryptedChatMessageDTO);
-			messageTemplate.convertAndSend("/sub/chat/room/" + chatMessageDTO.getChatRoomType().getType() + "/" + chatMessageDTO.getChatRoomId(), chatMessageDTO);
+			ChatMessageDTO chatMessageDTO = objectMapper.readValue(publishedMessage, ChatMessageDTO.class);
+			messageTemplate.convertAndSend(
+				"/sub/chat/room/" + chatMessageDTO.getChatRoomType().getType() + "/" + chatMessageDTO.getChatRoomId(),
+				chatMessageDTO);
 		} catch (Exception e) {
 			throw new ChatMessageSendingException("Failed to send chat message", CHAT_MESSAGE_SENDING_FAILED);
 		}
